Add files via upload
Through this analysis of the sentiment and topics within the YouTube comments, we intend to examine whether there is a discernible alignment between the sentiments expressed and the rise of far-right ideology. Furthermore, we aim to assess whether these sentiments are indicative of broader support for the anti-immigration protests, thereby shedding light on the potential growth of a far-right movement in Ireland.